+++ app/src/main/java/com/tautech/cclapp/daos/DriverDao.kt
@@ -9,7 +9,7 @@ interface DriverDao {
     fun getAll(): List<Driver>
 
     @Query("SELECT * FROM driver WHERE id IN (:driverIds)")
-    fun loadAllByIds(driverIds: IntArray): List<Driver>
+    fun loadAllByIds(driverIds: LongArray): List<Driver>
 
     @Query("SELECT * FROM driver WHERE id = CAST(:id AS NUMERIC)")
     fun getById(id: Long?): Driver
