+++ app/src/main/java/com/tautech/cclapp/models/Attributes.kt
@@ -3,7 +3,7 @@ package com.tautech.cclapp.models
 import java.io.Serializable
 
 data class Attributes(
-    var driverId: List<Int>? = null,
+    var driverId: List<Long>? = null,
     var userType: List<String>? = null): Serializable {
     
 }
